experiments/temporal-models/bbox-tube-temporal/tests/test_reproducibility.py
"""End-to-end reproducibility test for the training pipeline.

Runs two short Lightning fits with the same seed on a tiny fake dataset and
asserts that every weight in the final ``state_dict`` is bitwise identical.
A third run with a different seed acts as a negative control so the test
cannot silently pass if nothing is actually random.

Exercises the full seeding path used by ``scripts/train.py``:
``L.seed_everything(seed, workers=True)`` + ``Trainer(deterministic=True)``
with ``num_workers > 0`` so Lightning's auto-injected ``worker_init_fn`` is
in play.
"""

from __future__ import annotations

import json
from pathlib import Path

import lightning as L
import numpy as np
import torch
from PIL import Image
from torch.utils.data import DataLoader

from bbox_tube_temporal.dataset import TubePatchDataset
from bbox_tube_temporal.lit_temporal import LitTemporalClassifier

SEED = 1234
OTHER_SEED = 5678


def _make_split(
    root: Path, split_name: str, samples: list[tuple[str, int, int]]
) -> Path:
    split = root / split_name
    split.mkdir()
    index = []
    for seq_id, label_int, num_frames in samples:
        seq_dir = split / seq_id
        seq_dir.mkdir()
        for i in range(num_frames):
            arr = np.full((224, 224, 3), 50 + i * 5, dtype=np.uint8)
            Image.fromarray(arr).save(seq_dir / f"frame_{i:02d}.png")
        meta = {
            "sequence_id": seq_id,
            "split": split_name,
            "label": "smoke" if label_int == 1 else "fp",
            "label_int": label_int,
            "num_frames": num_frames,
            "context_factor": 1.5,
            "patch_size": 224,
            "frames": [
                {
                    "frame_idx": i,
                    "frame_id": f"f{i}",
                    "is_gap": False,
                    "orig_bbox": [0.5, 0.5, 0.05, 0.05],
                    "crop_bbox_pixels": [0, 0, 100, 100],
                    "filename": f"frame_{i:02d}.png",
                }
                for i in range(num_frames)
            ],
        }
        (seq_dir / "meta.json").write_text(json.dumps(meta))
        index.append(
            {"sequence_id": seq_id, "label_int": label_int, "num_frames": num_frames}
        )
    (split / "_index.json").write_text(json.dumps(index))
    return split


def _fit_once(seed: int, train_dir: Path, val_dir: Path, log_dir: Path) -> dict:
    L.seed_everything(seed, workers=True)

    train_ds = TubePatchDataset(train_dir, max_frames=5)
    val_ds = TubePatchDataset(val_dir, max_frames=5)
    train_loader = DataLoader(
        train_ds,
        batch_size=2,
        shuffle=True,
        num_workers=2,
        persistent_workers=True,
    )
    val_loader = DataLoader(
        val_ds,
        batch_size=2,
        shuffle=False,
        num_workers=2,
        persistent_workers=True,
    )

    lit = LitTemporalClassifier(
        backbone="resnet18",
        arch="gru",
        hidden_dim=16,
        learning_rate=1e-3,
        weight_decay=1e-2,
        pretrained=False,
        num_layers=1,
        bidirectional=False,
    )

    trainer = L.Trainer(
        max_epochs=2,
        accelerator="cpu",
        devices=1,
        deterministic=True,
        logger=False,
        enable_checkpointing=False,
        enable_progress_bar=False,
        enable_model_summary=False,
        log_every_n_steps=1,
        default_root_dir=log_dir,
    )
    trainer.fit(lit, train_loader, val_loader)

    return {k: v.detach().clone() for k, v in lit.state_dict().items()}


def test_training_is_bitwise_reproducible_with_fixed_seed(tmp_path: Path) -> None:
    train_dir = _make_split(
        tmp_path,
        "train",
        [("a", 1, 5), ("b", 0, 4), ("c", 1, 3), ("d", 0, 5)],
    )
    val_dir = _make_split(tmp_path, "val", [("e", 1, 4), ("f", 0, 3)])

    run1 = _fit_once(SEED, train_dir, val_dir, tmp_path / "run1")
    run2 = _fit_once(SEED, train_dir, val_dir, tmp_path / "run2")
    run_other = _fit_once(OTHER_SEED, train_dir, val_dir, tmp_path / "run_other")

    assert run1.keys() == run2.keys() == run_other.keys()

    for key in run1:
        assert torch.equal(run1[key], run2[key]), (
            f"Same-seed runs diverged at parameter {key!r} — training is not "
            f"reproducible. Max abs diff: "
            f"{(run1[key].float() - run2[key].float()).abs().max().item()}"
        )

    differing = [key for key in run1 if not torch.equal(run1[key], run_other[key])]
    assert differing, (
        "Different-seed run produced identical weights — the test dataset is "
        "too trivial to distinguish seeds, so same-seed equality does not prove "
        "reproducibility. Increase training work."
    )
