experiments/temporal-models/bbox-tube-temporal/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bbox_tube_temporal"]

[project]
name = "bbox-tube-temporal"
version = "0.1.0"
description = "Temporal smoke classifier over YOLO-detected tube patches (timm backbone + mean-pool/GRU head)"
requires-python = ">=3.11"
dependencies = [
    "pyrocore",
    "lightning>=2.2",
    "matplotlib>=3.8",
    "numpy>=1.26,<2",
    "opencv-python-headless>=4.8",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.4",
    "tensorboard>=2.16",
    "timm>=1.0",
    "torch>=2.2",
    "torchvision>=0.17",
    "tqdm>=4.66",
    "ultralytics>=8.3",
]

[tool.uv.sources]
pyrocore = { path = "../../../lib/pyrocore" }

[dependency-groups]
dev = [
    "dvc[s3]>=3.56",
    "jupyterlab>=4.0",
    "nbqa>=1.9",
    "nbstripout>=0.8",
    "pytest>=8.0",
    "ruff>=0.9",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM", "PLC0415"]

[tool.ruff.format]
quote-style = "double"
